+++ libs/btp/navigation/components/navigation-link/navigation-link.component.ts
@@ -11,6 +11,7 @@ import {
     Signal,
     TemplateRef,
     ViewEncapsulation,
+    booleanAttribute,
     inject
 } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
@@ -33,7 +34,7 @@ export class NavigationLinkRefDirective {
 
 @Component({
     // eslint-disable-next-line @angular-eslint/component-selector
-    selector: 'a[fdb-navigation-link]',
+    selector: 'a[fdb-navigation-link], button[fdb-navigation-link]',
     imports: [IconComponent],
     hostDirectives: [RouterLinkActive],
     changeDetection: ChangeDetectionStrategy.OnPush,
@@ -46,7 +47,7 @@ export class NavigationLinkRefDirective {
     ],
     host: {
         class: 'fd-navigation__link',
-        role: 'link'
+        '[attr.role]': 'quickCreate ? "button" : "link"'
     },
     templateUrl: './navigation-link.component.html'
 })
@@ -66,6 +67,10 @@ export class NavigationLinkComponent extends FdbNavigationItemLink implements On
     /** Whether the link is for the external resource. */
     @Input()
     external = false;
+
+    /** Whether it's a quick create button. */
+    @Input({ transform: booleanAttribute })
+    quickCreate = false;
 
     /** @hidden */
     @HostBinding('attr.tabindex')