+++ libs/core/list/directives/subline/list-subline.directive.spec.ts
@@ -1,0 +1,38 @@
+import { Component, ElementRef, ViewChild } from '@angular/core';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ListSublineDirective } from './list-subline.directive';
+
+@Component({
+    template: ` <li #componentElement fd-list-subline>List Subline Directive Test</li> `,
+    standalone: true,
+    imports: [ListSublineDirective]
+})
+class TestComponent {
+    @ViewChild('componentElement', { read: ElementRef })
+    ref: ElementRef;
+}
+
+describe('ListSublineDirective', () => {
+    let component: TestComponent;
+    let fixture: ComponentFixture<TestComponent>;
+
+    beforeEach(waitForAsync(() => {
+        TestBed.configureTestingModule({
+            imports: [TestComponent]
+        }).compileComponents();
+    }));
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(TestComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+
+    it('should assign class', () => {
+        expect(component.ref.nativeElement.className).toBe('fd-list__subline');
+    });
+});