+++ libs/core/settings/settings.component.spec.ts
@@ -1,0 +1,30 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { whenStable } from '@fundamental-ngx/core/tests';
+import { SettingsComponent } from './settings.component';
+
+describe('SettingsContainerComponent', () => {
+    let component: SettingsComponent;
+    let fixture: ComponentFixture<SettingsComponent>;
+
+    beforeEach(waitForAsync(() => {
+        TestBed.configureTestingModule({
+            imports: [SettingsComponent]
+        }).compileComponents();
+    }));
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(SettingsComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+
+    it('should apply proper css classes', async () => {
+        await whenStable(fixture);
+
+        expect(fixture.nativeElement.classList.contains('fd-settings')).toBe(true);
+    });
+});